Issue #33: Consolidate batch and micro-batch reader classes
